Single error message without list
This changes the behaviour of the error message when only one error was raised. Previously, a list (<ul>) would be generated with a single item in it.

I found this to be semantically strange and also visually ugly, hence this commit. Now only the error message itself is displayed when only one error was raised for the given field.